TISE solver using matrix numerov
These could be optimized in a couple of ways:
First, we could try a sparse solver for the eigenvalue problem and only solve for the first X eigenvalues, which we can estimate using the equation from Rosendahl 2022.
Second, we can probably vectorize out the for loops. 
However, these will do the job in about a second.